requirements.txt
anthropic>=0.49.0
streamlit>=1.42.0
pyyaml>=6.0
